Refactor `contracts` cmd
