--- src/commands/contracts/l1_shared_bridge/mod.rs
@@ -1,4 +1,0 @@
-use clap::Subcommand;
-
-#[derive(Subcommand, PartialEq)]
-pub(crate) enum Command {}
